pyproject.toml
[project]
name = "oprocess"
version = "0.1.0"
description = "AI-native process classification MCP Server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0",
    "sqlite-vec>=0.1.6",
]

[project.optional-dependencies]
embed = ["google-genai>=1.0"]
dev = [
    "pytest>=7.0",
    "pytest-benchmark>=4.0",
    "pytest-cov>=4.0",
    "ruff>=0.5",
    "google-genai>=1.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "--cov=oprocess",
    "--cov-report=term-missing:skip-covered",
    "--cov-branch",
    "--cov-fail-under=80",
    "--benchmark-skip",
    "--ignore=tests/test_performance.py",
]

[tool.coverage.run]
source = ["src/oprocess"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/oprocess"]

[tool.oprocess]
boundary_threshold = 0.45
audit_log_enabled = true
default_language = "zh"
